+++ cooked_input/__init__.py
@@ -19,7 +19,7 @@ from .error_callbacks import MaxRetriesError, ConvertorError, ValidationError
 from .error_callbacks import print_error, log_error, silent_error, DEFAULT_CONVERTOR_ERROR, DEFAULT_VALIDATOR_ERROR
 from .convertors import TABLE_ID, TABLE_VALUE, TABLE_ID_OR_VALUE
 from .convertors import Convertor, IntConvertor, FloatConvertor, BooleanConvertor
-from .convertors import ListConvertor, DateConvertor, YesNoConvertor, TableConvertor, ChoiceConvertor
+from .convertors import ListConvertor, DateConvertor, YesNoConvertor, ChoiceConvertor
 from .validators import Validator, LengthValidator, EqualToValidator, RangeValidator
 from .validators import AnyOfValidator, NoneOfValidator, ChoiceValidator, RegexValidator, PasswordValidator
 from .validators import IsFileValidator, ListValidator, SimpleValidator